pyproject.toml
[project]
name = "biorag"
version = "0.1.0"
description = "Biological PDF RAG System for scientific literature querying"
requires-python = ">=3.9"
dependencies = [
    "dspy-ai",
    "faiss-cpu",
    "openai",
    "pypdf",
    "pymupdf",
    "tiktoken",
    "python-dotenv",
    "jsonlines",
    "tqdm",
    "numpy",
    "pandas",
    "requests>=2.32.3",
    "httpx>=0.28.1",
    "aiohttp>=3.12.7",
    "biopython>=1.85",
    "mygene>=3.2.2",
    "reactome2py>=3.0.0",
    "tenacity>=9.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"] 

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.11.12",
]
